+++ routes/web.php
@@ -27,7 +27,7 @@ Route::prefix('admin')->group(function() {
         * Manage user
         */
        Route::resource('user','ManageUserController', ['as' => 'admin']);
-       Route::put('user/{id}/update-image', 'ManageUserController@updateImage')->name('admin.user.update.image');
+       Route::put('user/{id}/update-image', 'UserProfileController@updateImage')->name('admin.user.update.image');
    });
 });
